+++ src/Api/WpCacheTrait.php
@@ -10,31 +10,21 @@ namespace TheFrosty\WpUtilities\Api;
 trait WpCacheTrait
 {
 
+    use Hash;
+
     /**
      * Cache group value.
      *
-     * @var string $group
+     * @var string|null $group
      */
-    private $group = __CLASS__;
-
-    /**
-     * Get a cache key.
-     *
-     * @param string $data
-     *
-     * @return string
-     */
-    protected function getHashedKey(string $data): string
-    {
-        return \hash('sha256', $data);
-    }
+    private $group;
 
     /**
      * Get the cache group.
      */
     protected function getCacheGroup(): string
     {
-        return $this->group;
+        return $this->group ?? static::class;
     }
 
     /**
@@ -62,7 +52,7 @@ trait WpCacheTrait
      */
     protected function getCache(string $key, ?string $group = null, bool $force = false, ?bool &$found = null)
     {
-        return \wp_cache_get($key, $group ?? $this->group, $force, $found);
+        return \wp_cache_get($key, $group ?? $this->getCacheGroup(), $force, $found);
     }
 
     /**
@@ -75,8 +65,21 @@ trait WpCacheTrait
      *
      * @return bool Returns TRUE on success or FALSE on failure.
      */
-    protected function setCache(string $key, $value, ?string $group = '', int $expiration = 0): bool
+    protected function setCache(string $key, $value, ?string $group = null, int $expiration = 0): bool
     {
-        return \wp_cache_set($key, $value, $group ?? $this->group, $expiration);
+        return \wp_cache_set($key, $value, $group ?? $this->getCacheGroup(), $expiration);
+    }
+
+    /**
+     * Deletes a value in cache.
+     *
+     * @param string $key The key under which the value is stored.
+     * @param string|null $group The group value appended to the $key.
+     *
+     * @return bool Returns TRUE on success or FALSE on failure.
+     */
+    protected function deleteCache(string $key, ?string $group = null): bool
+    {
+        return \wp_cache_delete($key, $group ?? $this->getCacheGroup());
     }
 }